Ajustes de migrations do dominio do lojista e docker-compose; remove RARs legados
Corrige tipos/constraints/FKs nas migrations V002-V023 (control-plane e
dominio lojista), ajusta docker-compose e remove os dumps legados
db/*.RAR (ja preservados fora do controle de versao).

+++ db/migration/V002__contexto_tenant.sql
@@ -11,11 +11,11 @@
 -- As tabelas de PLATAFORMA (este schema) NÃO usam esta função (são globais, P9).
 
 CREATE OR REPLACE FUNCTION plataforma.tenant_atual()
-RETURNS bigint
+RETURNS smallint
 LANGUAGE sql
 STABLE
 AS $$
-  SELECT NULLIF(current_setting('app.id_tenant', true), '')::bigint;
+  SELECT NULLIF(current_setting('app.id_tenant', true), '')::smallint;
 $$;
 
 COMMENT ON FUNCTION plataforma.tenant_atual() IS
